+++ docs/source/changelog.md
@@ -28,6 +28,7 @@ Release notes for the `space_packet_parser` library
 - Add function to directly create an `xarray.DataSet` from a packet file and XTCE definition.
   e.g. `space_packet_parser.xarr.create_dataset([packets1, packets2, ...], definition)`
 - BUGFIX: update list of allowed float encodings to match XTCE spec
+- Add benchmark tests and documentation overview of benchmarks.
 
 ### v5.0.1 (released)
 - BUGFIX: Allow raw_value representation for enums with falsy raw values. Previously these defaulted to the enum label.